backend/requirements.txt
Flask
flask-cors
Flask-SQLAlchemy
python-dotenv
pymysql
mysql-connector-python
